requirements.txt
keyboard==0.13.5
pywin32==306
pystray==0.19.4
Pillow==10.3.0
pyinstaller==6.8.0 